+++ packages/form/src/components/FormItem/index.tsx
@@ -105,6 +105,16 @@ type WarpFormItemProps = {
   addonAfter?: React.ReactNode;
   /**
    * @name 获取时转化值，一般用于将数据格式化为组件接收的格式
+   * @param value 字段的值
+   * @param namePath 字段的name
+   * @returns 字段新的值
+   *
+   *
+   * @example a,b => [a,b]     convertValue: (value,namePath)=> value.split(",")
+   * @example string => json   convertValue: (value,namePath)=> JSON.parse(value)
+   * @example number => date   convertValue: (value,namePath)=> Moment(value)
+   * @example YYYY-MM-DD => date   convertValue: (value,namePath)=> Moment(value,"YYYY-MM-DD")
+   * @example  string => object   convertValue: (value,namePath)=> { return {value,label:value} }
    */
   convertValue?: SearchConvertKeyFn;
 };
@@ -199,12 +209,24 @@ const WarpFormItem: React.FC<FormItemProps & WarpFormItemProps> = ({
   );
 };
 
-type ProFormItemProps = FormItemProps & {
+export type ProFormItemProps = FormItemProps & {
   ignoreFormItem?: boolean;
   valueType?: ProSchemaValueType<'text'>;
-  /** @name 提交时转化值，一般用于将值转化为提交的数据 */
+  /**
+   * @name 提交时转化值，一般用于将值转化为提交的数据
+   * @param value 字段的值
+   * @param namePath 字段的name
+   * @param allValues 所有的字段
+   * @returns 字段新的值，如果返回对象，会和所有值 merge 一次
+   *
+   * @example {name:[a,b] => {name:a,b }    transform: (value,namePath,allValues)=> value.join(",")
+   * @example {name: string => { newName:string }    transform: (value,namePath,allValues)=> { newName:value }
+   * @example {name:moment} => {name:string transform: (value,namePath,allValues)=> value.format("YYYY-MM-DD")
+   * @example {name:moment}=> {name:时间戳} transform: (value,namePath,allValues)=> value.valueOf()
+   * @example {name:{value,label}} => { name:string} transform: (value,namePath,allValues)=> value.value
+   * @example {name:{value,label}} => { valueName,labelName  } transform: (value,namePath,allValues)=> { valueName:value.value, labelName:value.name }
+   */
   transform?: SearchTransformKeyFn;
-
   dataFormat?: string;
   lightProps?: LightWrapperProps;
   proFormFieldKey?: any;
@@ -244,7 +266,7 @@ const ProFormItem: React.FC<ProFormItemProps> = (props) => {
     // Field.type === 'ProField' 时 props 里面是有 valueType 的，所以要设置一下
     // 写一个 ts 比较麻烦，用 any 顶一下
     setFieldValueType(
-      [formListField.listName, name].flat(1).filter((itemName) => itemName !== undefined),
+      [formListField.listName, props.name].flat(1).filter((itemName) => itemName !== undefined),
       {
         valueType: valueType || 'text',
         dateFormat: dataFormat,
@@ -314,6 +336,5 @@ const ProFormItem: React.FC<ProFormItemProps> = (props) => {
   );
 };
 
-export type { ProFormItemProps };
 export { FormItemProvide };
 export default ProFormItem;