+++ packages/form/src/interface.ts
@@ -1,4 +1,4 @@
-import type { FormItemProps, SpaceProps } from 'antd';
+import type { FormItemProps, RowProps, SpaceProps } from 'antd';
 import type { ProFormItemProps } from './components';
 import type { ProFormInstance } from './BaseForm';
 import type { LabelTooltipType } from 'antd/lib/form/FormItemLabel';
@@ -8,8 +8,30 @@ import type {
   ProFieldValueType,
   ProSchema,
   SearchConvertKeyFn,
-  SearchTransformKeyFn,
 } from '@ant-design/pro-utils';
+import type { ColProps } from 'antd/lib/grid/col';
+
+export interface ProFormGridConfig {
+  /**
+   * open grid layout
+   * @default false
+   */
+  grid?: boolean;
+  /**
+   * only works when grid is enabled
+   *
+   * When passing the `span` attribute, the default value is empty
+   * @default
+   * { xs: 24 }
+   */
+  colProps?: ColProps;
+  /**
+   * only works when grid is enabled
+   * @default
+   * { gutter: 8 }
+   */
+  rowProps?: RowProps;
+}
 
 export type ProFormItemCreateConfig = {
   /** 自定义类型 */
@@ -46,9 +68,6 @@ export type ExtendsProps = {
    * @name 只读模式
    */
   readonly?: boolean;
-
-  /** @name 提交时转化值，一般用于数组类型 */
-  transform?: SearchTransformKeyFn;
 
   /**
    * @name 获取时转化值，一般用于将数据格式化为组件接收的格式
@@ -91,7 +110,7 @@ export type GroupProps = {
   onCollapse?: (collapsed: boolean) => void;
   /** 自定选中一个input，只能有一个生效 */
   autoFocus?: boolean;
-};
+} & ProFormGridConfig;
 
 export type FieldProps = {
   style?: React.CSSProperties;
@@ -127,6 +146,7 @@ export type ProFormFieldItemProps<T = Record<string, any>> = {
   /** QueryFilter 上的footer */
   footerRender?: LightFilterFooterRender;
 } & Omit<ProFormItemProps, 'valueType'> &
+  Pick<ProFormGridConfig, 'colProps'> &
   ExtendsProps;
 
 /**