+++ packages/table/src/demos/single.tsx
@@ -43,6 +43,7 @@ const columns: ProColumns<GithubIssueItem>[] = [
     },
   },
   {
+    disable: true,
     title: '状态',
     dataIndex: 'state',
     filters: true,
@@ -66,6 +67,7 @@ const columns: ProColumns<GithubIssueItem>[] = [
     },
   },
   {
+    disable: true,
     title: '标签',
     dataIndex: 'labels',
     search: false,
@@ -107,6 +109,7 @@ const columns: ProColumns<GithubIssueItem>[] = [
   {
     title: '操作',
     valueType: 'option',
+    key: 'option',
     render: (text, record, _, action) => [
       <a
         key="editable"
@@ -160,6 +163,9 @@ export default () => {
       columnsState={{
         persistenceKey: 'pro-table-singe-demos',
         persistenceType: 'localStorage',
+        onChange(value) {
+          console.log('value: ', value);
+        },
       }}
       rowKey="id"
       search={{